ff/serving/requirements.txt
grpcio==1.42.0
grpcio-tools==1.42.0
protobuf==3.19.1
six==1.16.0
